+++ lib/src/components/top_menu_bar.dart
@@ -12,8 +12,8 @@ import 'package:zego_uikit_prebuilt_video_conference/src/components/icon_defines
 import 'package:zego_uikit_prebuilt_video_conference/src/components/member/member_list_button.dart';
 import 'package:zego_uikit_prebuilt_video_conference/src/components/message/in_room_message_button.dart';
 import 'package:zego_uikit_prebuilt_video_conference/src/components/pop_up_manager.dart';
-import 'package:zego_uikit_prebuilt_video_conference/src/prebuilt_video_conference_config.dart';
-import 'package:zego_uikit_prebuilt_video_conference/src/prebuilt_video_conference_defines.dart';
+import 'package:zego_uikit_prebuilt_video_conference/src/config.dart';
+import 'package:zego_uikit_prebuilt_video_conference/src/defines.dart';
 
 /// @nodoc
 class ZegoTopMenuBar extends StatefulWidget {