+++ lib/zego_uikit_prebuilt_video_conference.dart
@@ -2,7 +2,7 @@ library zego_uikit_prebuilt_video_conference;
 
 export 'package:zego_uikit/zego_uikit.dart';
 
-export 'src/prebuilt_video_conference.dart';
-export 'src/prebuilt_video_conference_config.dart';
-export 'src/prebuilt_video_conference_controller.dart';
-export 'src/prebuilt_video_conference_defines.dart';
+export 'src/config.dart';
+export 'src/controller.dart';
+export 'src/defines.dart';
+export 'src/video_conference.dart';