+++ CHANGELOG.md
@@ -1,62 +1,103 @@
+
+## 2.6.3
+
+- Update dependencies
+
+## 2.6.2
+
+- Optimization warnings from analysis
+
+## 2.6.1
+
+- Optimization warnings from analysis
+
+## 2.6.0
+
+- Support listening for errors signaling plugins and uikit library.
+
+## 2.5.0
+
+- Support remove user and mute user by controller, which can be called through **ZegoUIKitPrebuiltVideoConferenceController.room**.
+
 ## 2.4.2
+
 - update dart dependency
 
 ## 2.4.1
+
 - remove http library dependency.
 
 ## 2.4.0
+
 - The bottom menu bar and top menu bar support custom background colors.
 
 ## 2.3.2
+
 - Update ReadME
 
 ## 2.3.1
+
 - Update dependencies
 
 ## 2.3.0
+
 - Added logic for being kicked out of the video conference, which will automatically exit and return to the previous page.
 
 ## 2.2.10
+
 - Fixed an issue where the avatar builder was not working for the member list.
 
 ## 2.2.9
+
 - fix the issue of conflict with extension key of the `flutter_screenutil` package.
 
 ## 2.2.8
+
 - Update dependencies
 
 ## 2.2.7
+
 - Update comments
 
 ## 2.2.6
+
 - deprecate flutter_screenutil_zego package
 
 ## 2.2.5
+
 - fix import issues
 
 ## 2.2.4
+
 - Update dependencies
 
 ## 2.2.3
+
 - Fix some issues
 
 ## 2.2.2
+
 - mark 'appDesignSize' as Deprecated
 
 ## 2.2.1
+
 - Update dependencies
 
 ## 2.2.0
+
 - To differentiate the 'appDesignSize' between the App and ZegoUIKitPrebuiltVideoConference, we introduced the 'flutter_screenutil_zego' library and removed the 'appDesignSize' parameter from the
   ZegoUIKitPrebuiltVideoConference that was previously present.
 
 ## 2.1.4
+
 - fixed appDesignSize for ScreenUtil that didn't work
 
 ## 2.1.3
+
 - add assert to key parameters to ensure prebuilt run normally
 
 ## 2.1.2
+
 - fixed landscape not displaying full web screen sharing content
 
 ## 2.1.1
@@ -69,6 +110,7 @@
 - update dependency
 
 ## 2.0.1
+
 - remove login token
 - optimizing code warnings
 
@@ -78,66 +120,66 @@
 
 ## 1.1.13
 
-* downgrade flutter_screenutil to ">=5.5.3+2 <5.6.1"
+- Downgrade flutter_screenutil to ">=5.5.3+2 <5.6.1"
 
 ## 1.1.12
 
-* update a dependency to the latest release
-* support sdk log
+- Update a dependency to the latest release
+- Support sdk log
 
 ## 1.1.11
 
-* support view screen sharing stream of web
+- Support view screen sharing stream of web
 
 ## 1.1.10
 
-* update a dependency to the latest release
+- Update a dependency to the latest release
 
 ## 1.1.9
 
-* update a dependency to the latest release
+- Update a dependency to the latest release
 
 ## 1.1.8
 
-* update a dependency to the latest release
+- Update a dependency to the latest release
 
 ## 1.1.7
 
-* update a dependency to the latest release
+- Update a dependency to the latest release
 
 ## 1.1.6
 
-* fix some bugs
-* update a dependency to the latest release
+- Fix some bugs
+- Update a dependency to the latest release
 
 ## 1.1.5
 
-* fix some bugs
+- Fix some bugs
 
 ## 1.1.4
 
-* fix some bugs
+- Fix some bugs
 
 ## 1.1.3
 
-* update a dependency to the latest release
+- Update a dependency to the latest release
 
 ## 1.1.2
 
-* fix some bugs
+- Fix some bugs
 
 ## 1.1.1
 
-* fix some bugs
+- Fix some bugs
 
 ## 1.1.0
 
-* support chat and notification.
+- Support chat and notification.
 
 ## 1.0.0
 
-* Congratulations!
+- Congratulations!
 
 ## 0.0.1
 
-* upload initial release.
+- Upload initial release.